Fixed handling of constants when the enclosing module isn’t the lexically enclosing module
When the enclosing module was created with Module.new, it isn’t inserted into the lexical lookup chain, and therefore isn’t the lexically enclosing module. The existing implementation handled that well enough to set any resulting constants in the correct place, but wasn’t putting the const_missing handler in the correct place. In this case, the “correct place” is Object.ancestors.first because Object.ancestors is the fallback lookup chain when Module.nesting is empty which is what happens when the enclosing module was created with Module.new.

This has the nice side-effect of cleaning up the tests.