Changed DSL recursion wrapper handling
When building a recursive grammar object that will ultimately be assigned to a constant, assign the recursive-grammar object rather than the Recursion proxy. This makes the resulting parse tree a bit more sensible, and makes it easier to use the resulting constant.

Note that while inside the DSL-block for the recursive grammar element the constant evaluates to the recursion proxy object rather than the not-yet-completed grammar element.