Fixed handling of an odd case of recursion
The Python grammar ends up making use of the recursion wrapper for a part of the grammar that isn't actually recursive. I suspect that I've simply translated the grammar improperly, or I'm missing a needed refactoring, but I haven't been able to figure it out, yet. Nor have I looked extensively. Regardless, it seems reasonable to consider this a bug in the recursion-handling.